:sparkles: 스케줄 생성 테스트 확인 완료
