+++ app/modules/schedule/permissions.py
@@ -1,0 +1,9 @@
+from app.modules.auth.models import User, PositionEnum
+
+
+def is_supervisor(user: User) -> bool:
+    return user.position in {
+        PositionEnum.advisor,
+        PositionEnum.assistant_manager,
+        PositionEnum.manager,
+    }
